a massive ridiculous amount of mypy changes

+++ src/ocrmypdf/builtin_plugins/tesseract_ocr.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 import argparse
 import logging
 import os
-from typing import Annotated
+from typing import Annotated, Any, Callable, cast
 
 from PIL import Image
 from pydantic import BaseModel, Field, field_validator, model_validator
@@ -168,7 +168,7 @@ class TesseractOptions(BaseModel):
         tess.add_argument(
             f'--{namespace}-timeout',
             default=180.0,
-            type=numeric(float, 0),
+            type=numeric(int, 0),
             metavar='SECONDS',
             dest=f'{namespace}_timeout',
             help=(
@@ -183,7 +183,7 @@ class TesseractOptions(BaseModel):
         tess.add_argument(
             f'--{namespace}-non-ocr-timeout',
             default=180.0,
-            type=numeric(float, 0),
+            type=numeric(int, 0),
             metavar='SECONDS',
             dest=f'{namespace}_non_ocr_timeout',
             help=(
